requirements.txt
# Core Python packages
PyQt6>=6.4.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0

# Machine Learning
scikit-learn>=1.1.0
tensorflow>=2.10.0
torch>=1.12.0

# Financial Data and Analysis
yfinance>=0.1.87
pandas-ta>=0.3.14b0
ta-lib>=0.4.24
alpaca-py>=0.8.0

# Data Visualization
matplotlib>=3.5.0
plotly>=5.10.0

# Database and Storage
sqlite3

# Security and Configuration
python-dotenv>=0.19.0
cryptography>=3.4.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2022.1 